+++ crates/cli/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- use `miette` for error-printing
+
+### Breaking changes
+
+- all new error types, cli no longer uses (formerly named) `contender_core::ContenderError`
+    - the cli now uses `contender_cli::error::CliError`
+
 ## [0.5.6](https://github.com/flashbots/contender/releases/tag/v0.5.6) - 2025-10-10
 
 - erc20: fuzz recipient address ([#366](https://github.com/flashbots/contender/pull/366))